Move mineBlocks and GenerateTickets to helper function
